// client/src/models/IMotherboard.ts
export interface IMotherboard {
    id: number;
    name: string;
    brand: string;
    socket: string;
    chipset: string;
    platform: string;
    form_factor: string;
    memory_type: string;
    number_of_slots: number;
    maximum_volume: number;
    maximum_frequency: number;
    vga: number;
    hdmi: number;
    dvi: number;
    display_port: number;
    lan: number;
    wifi: number;
    bluetooth: number;
    m2: number;
    sata: number;
    raid: number;
    price: number;
    img: string;
}